pyproject.toml
[project]
name = "nanogpt-analysis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
]
